examples/python/fastapi-server/pyproject.toml
[project]
name = "fastapi-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "fastapi>=0.120.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "openlibx402-core",
    "openlibx402-fastapi",
]
